requirements_q1.txt
# Q1 Publication-Grade Requirements for PI-HMARL
# Core requirements for rigorous experimental evaluation

# Core Scientific Computing
numpy>=1.19.0
scipy>=1.5.0
pandas>=1.1.0

# Deep Learning
torch>=1.7.0
torchvision>=0.8.0

# Statistical Analysis
scikit-learn>=0.24.0
statsmodels>=0.12.0
pymc3>=3.11.0  # For Bayesian analysis
arviz>=0.11.0  # For Bayesian visualization

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Configuration and Utils
pyyaml>=5.4.0
tqdm>=4.60.0
tabulate>=0.8.0

# Profiling and Performance
memory-profiler>=0.58.0
psutil>=5.8.0
py-cpuinfo>=8.0.0
gputil>=1.4.0  # GPU monitoring

# Symbolic Math (for theoretical analysis)
sympy>=1.7.0

# Additional utilities
h5py>=3.1.0  # For large data storage
joblib>=1.0.0  # For parallel processing
dill>=0.3.3  # For advanced serialization

# Optional but recommended for Q1 standards
# networkx>=2.5  # For graph analysis
# tensorboard>=2.4.0  # For experiment tracking